refactor: move Feature Flags source resolution to settings
